add archetype scraper
